+++ src/pidStore/internalPidStore.js
@@ -1,0 +1,5 @@
+'use strict';
+
+// This file exists to avoid dependency cycles between:
+//   logger => agent/bunyanToAgentStream => pidStore/index => logger
+exports.pid = process.pid;